Refactor. Create new AvailabilityStateService for source of truth for building the state model. Move 'GetBooking' methods from BookingService into a new BookingQueryService to prevent circular references and isolate logic.
