+++ src/api/Nhs.Appointments.Core/IAvailabilityService.cs
@@ -9,7 +9,6 @@ public interface IAvailabilityService
     Task<IEnumerable<AvailabilityCreatedEvent>> GetAvailabilityCreatedEventsAsync(string site, DateOnly from);
     Task<IEnumerable<DailyAvailability>> GetDailyAvailability(string site, DateOnly from, DateOnly to);
     Task CancelSession(string site, DateOnly date, string from, string until, string[] services, int slotLength, int capacity);
-    Task<AvailabilityState> GetAvailabilityState(string site, DateTime from, DateTime to, string service, bool processRecalculations);
     Task<AvailabilityState> RecalculateAppointmentStatuses(string site, DateOnly day);
     Task<(bool Success, string Reference)> MakeBooking(Booking booking);
     Task<BookingCancellationResult> CancelBooking(string bookingReference, string site);
