+++ tests/Nhs.Appointments.Core.UnitTests/AvailabilityStateService/AvailabilityStateServiceTestBase.cs
@@ -1,29 +1,16 @@
-﻿using Nhs.Appointments.Core.Concurrency;
-using Nhs.Appointments.Core.Messaging;
+﻿namespace Nhs.Appointments.Core.UnitTests.AvailabilityStateService;
 
-namespace Nhs.Appointments.Core.UnitTests.AvailabilityCalculations;
-
-[MockedFeatureToggle("MultipleServicesEnabled", true)]
-public class AvailabilityCalculationsBase : FeatureToggledTests
+public class AvailabilityStateServiceTestBase
 {
-    protected readonly AvailabilityService _sut;
+    protected Core.AvailabilityStateService _sut;
     protected readonly Mock<TimeProvider> _timeProvider = new();
     private readonly Mock<IAvailabilityStore> _availabilityStore = new();
-    private readonly Mock<IAvailabilityCreatedEventStore> _availabilityCreatedEventStore = new();
-    private readonly Mock<IBookingsService> _bookingsService = new();
-    private readonly Mock<ISiteLeaseManager> _siteLeaseManager = new();
     private readonly Mock<IBookingsDocumentStore> _bookingsDocumentStore = new();
-    private readonly Mock<IReferenceNumberProvider> _referenceNumberProvider = new();
-    private readonly Mock<IBookingEventFactory> _eventFactory = new();
-    private readonly Mock<IMessageBus> _messageBus = new();
     
     protected const string MockSite = "some-site";
 
-    protected AvailabilityCalculationsBase() : base(typeof(AvailabilityCalculationsBase)) => _sut =
-        new AvailabilityService(_availabilityStore.Object,
-        _availabilityCreatedEventStore.Object, _bookingsService.Object, _siteLeaseManager.Object,
-        _bookingsDocumentStore.Object, _referenceNumberProvider.Object, _eventFactory.Object, _messageBus.Object,
-        _timeProvider.Object, _featureToggleHelper.Object);
+    protected AvailabilityStateServiceTestBase() => _sut =
+        new Core.AvailabilityStateService(_availabilityStore.Object, new BookingQueryService(_bookingsDocumentStore.Object, _timeProvider.Object));
 
     private DateTime TestDateAt(string time)
     {
@@ -53,8 +40,7 @@ public class AvailabilityCalculationsBase : FeatureToggledTests
 
     protected void SetupAvailabilityAndBookings(List<Booking> bookings, List<SessionInstance> sessions)
     {
-        _bookingsService
-            .Setup(x => x.GetBookings(It.IsAny<DateTime>(), It.IsAny<DateTime>(), MockSite, It.IsAny<string>()))
+        _bookingsDocumentStore.Setup(x => x.GetInDateRangeAsync(It.IsAny<DateTime>(), It.IsAny<DateTime>(), MockSite, It.IsAny<string>()))
             .ReturnsAsync(bookings);
 
         _availabilityStore
