+++ tests/Nhs.Appointments.Core.UnitTests/AvailabilityStateService/SingleServiceTests.cs
@@ -1,6 +1,6 @@
-﻿namespace Nhs.Appointments.Core.UnitTests.AvailabilityCalculations;
+﻿namespace Nhs.Appointments.Core.UnitTests.AvailabilityStateService;
 
-public class SingleServiceTests : AvailabilityCalculationsBase
+public class SingleServiceTests : AvailabilityStateServiceTestBase
 {
     [Fact]
     public async Task MakesNoChangesIfAllAppointmentsAreStillValid()
@@ -16,7 +16,7 @@ public class SingleServiceTests : AvailabilityCalculationsBase
 
         SetupAvailabilityAndBookings(bookings, sessions);
 
-        var resultingAvailabilityState = await _sut.GetAvailabilityState(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
+        var resultingAvailabilityState = await _sut.Build(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
 
         resultingAvailabilityState.Recalculations.Should().BeEmpty();
         resultingAvailabilityState.Bookings.Should().BeEquivalentTo(bookings);
@@ -37,7 +37,7 @@ public class SingleServiceTests : AvailabilityCalculationsBase
 
         SetupAvailabilityAndBookings(bookings, sessions);
 
-        var resultingAvailabilityState = await _sut.GetAvailabilityState(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
+        var resultingAvailabilityState = await _sut.Build(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
 
         resultingAvailabilityState.Recalculations.Should().HaveCount(3);
         resultingAvailabilityState.Bookings.Should().BeEquivalentTo(bookings);
@@ -57,7 +57,7 @@ public class SingleServiceTests : AvailabilityCalculationsBase
 
         SetupAvailabilityAndBookings(bookings, sessions);
 
-        var resultingAvailabilityState = await _sut.GetAvailabilityState(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
+        var resultingAvailabilityState = await _sut.Build(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
 
         resultingAvailabilityState.Recalculations.Should().ContainSingle(s =>
             s.Booking.Reference == "2" && s.Action == AvailabilityUpdateAction.SetToOrphaned);
@@ -81,7 +81,7 @@ public class SingleServiceTests : AvailabilityCalculationsBase
 
         SetupAvailabilityAndBookings(bookings, sessions);
 
-        var resultingAvailabilityState = await _sut.GetAvailabilityState(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 10, 0), "Green");
+        var resultingAvailabilityState = await _sut.Build(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 10, 0), "Green");
 
         resultingAvailabilityState.Recalculations.Should().Contain(s =>
             s.Booking.Reference == "1" && s.Action == AvailabilityUpdateAction.SetToOrphaned);
@@ -107,7 +107,7 @@ public class SingleServiceTests : AvailabilityCalculationsBase
 
         SetupAvailabilityAndBookings(bookings, sessions);
 
-        var resultingAvailabilityState = await _sut.GetAvailabilityState(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 10, 0), "Green");
+        var resultingAvailabilityState = await _sut.Build(MockSite, new DateTime(2025, 1, 1, 9, 0, 0), new DateTime(2025, 1, 1, 9, 10, 0), "Green");
 
         resultingAvailabilityState.Recalculations.Should().Contain(s =>
             s.Booking.Reference == "1" && s.Action == AvailabilityUpdateAction.SetToOrphaned);
@@ -134,7 +134,7 @@ public class SingleServiceTests : AvailabilityCalculationsBase
 
         SetupAvailabilityAndBookings(bookings, sessions);
 
-        var resultingAvailabilityState = await _sut.GetAvailabilityState(MockSite, new DateTime(2025, 1, 1, 9, 30, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
+        var resultingAvailabilityState = await _sut.Build(MockSite, new DateTime(2025, 1, 1, 9, 30, 0), new DateTime(2025, 1, 1, 9, 30, 0), "Green");
 
         resultingAvailabilityState.Bookings.Should().ContainSingle(b => b.Reference == "2");
         resultingAvailabilityState.Recalculations.Should().ContainSingle(r =>
